fix: long por string en todas las consultas

+++ sicf-consultas-generales/src/main/java/ob/unibanca/sicf/consultasgenerales/model/criterio/incomingvisa/CriterioBusquedaIncomingVisa.java
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class CriterioBusquedaIncomingVisa {
-	Integer idSecuenciaIncoming;
+	Long idSecuenciaIncoming;
 	@DateTimeFormat(pattern = "dd/MM/yyyy")
 	Date fechaProceso;
 }
